switch new project add to create method
